Refactoring placement of icons
