+++ src/editors/ied/access-point-container.ts
@@ -12,7 +12,7 @@ import './server-container.js'
 import { nothing } from 'lit-html';
 import { getDescriptionAttribute, getNameAttribute } from '../../foundation.js';
 import { Nsdoc } from '../../foundation/nsdoc.js';
-import { accessPointIcon } from './icons.js';
+import { accessPointIcon } from '../../icons/ied-icons.js';
 
 /** [[`IED`]] plugin subeditor for editing `AccessPoint` element. */
 @customElement('access-point-container')
