+++ src/editors/ied/ied-container.ts
@@ -14,7 +14,6 @@ import '../../action-pane.js';
 import './access-point-container.js';
 import { Nsdoc } from '../../foundation/nsdoc.js';
 import { getDescriptionAttribute, getNameAttribute, newWizardEvent } from '../../foundation.js';
-import { iedIcon } from './icons.js';
 
 /** [[`IED`]] plugin subeditor for editing `IED` element. */
 @customElement('ied-container')
@@ -39,7 +38,7 @@ export class IedContainer extends LitElement {
   }
 
   render(): TemplateResult {
-    return html`<action-pane .icon=${iedIcon} .label="${this.header()}">
+    return html`<action-pane icon="developer_board" .label="${this.header()}">
       <abbr slot="action" title="${translate('edit')}">
         <mwc-icon-button
           icon="edit"
