+++ src/wizards/controlwithiedname.ts
@@ -18,7 +18,7 @@ import {
   WizardInput,
 } from '../foundation.js';
 
-import { inputIcon } from '../icons.js';
+import { inputIcon } from '../icons/icons.js';
 import {
   getSourceReferences,
   openCommunicationMappingWizard,
