+++ src/zeroline/ied-editor.ts
@@ -13,7 +13,7 @@ import { Fab } from '@material/mwc-fab';
 
 import '../action-icon.js';
 import { createClientLnWizard } from '../wizards/clientln.js';
-import { gooseIcon, smvIcon, reportIcon } from '../icons.js';
+import { gooseIcon, smvIcon, reportIcon } from '../icons/icons.js';
 import { wizards } from '../wizards/wizard-library.js';
 import { newWizardEvent } from '../foundation.js';
 import { selectGseControlWizard } from '../wizards/gsecontrol.js';
